cleanup some imports and separate native sample traits
