+++ src/block/samples.rs
@@ -109,3 +109,84 @@ impl From<Sample> for f16 { #[inline] fn from(s: Sample) -> Self { s.to_f16() } 
 impl From<Sample> for f32 { #[inline] fn from(s: Sample) -> Self { s.to_f32() } }
 impl From<Sample> for u32 { #[inline] fn from(s: Sample) -> Self { s.to_u32() } }
 
+
+/// Create an arbitrary sample type from one of the defined sample types.
+/// Should be compiled to a no-op where the file contains the predicted sample type.
+pub trait FromNativeSample: Sized + Copy + Default + 'static {
+
+    /// Create this sample from a f16, trying to represent the same numerical value
+    fn from_f16(value: f16) -> Self;
+
+    /// Create this sample from a f32, trying to represent the same numerical value
+    fn from_f32(value: f32) -> Self;
+
+    /// Create this sample from a u32, trying to represent the same numerical value
+    fn from_u32(value: u32) -> Self;
+}
+
+// TODO haven't i implemented this exact behaviour already somewhere else in this library...??
+impl FromNativeSample for f32 {
+    fn from_f16(value: f16) -> Self { value.to_f32() }
+    fn from_f32(value: f32) -> Self { value } // this branch means that we never have to match every single sample if the file format matches the expected output
+    fn from_u32(value: u32) -> Self { value as f32 }
+}
+
+impl FromNativeSample for u32 {
+    fn from_f16(value: f16) -> Self { value.to_f32() as u32 }
+    fn from_f32(value: f32) -> Self { value as u32 }
+    fn from_u32(value: u32) -> Self { value }
+}
+
+impl FromNativeSample for f16 {
+    fn from_f16(value: f16) -> Self { value }
+    fn from_f32(value: f32) -> Self { f16::from_f32(value) }
+    fn from_u32(value: u32) -> Self { f16::from_f32(value as f32) }
+}
+
+impl FromNativeSample for Sample {
+    fn from_f16(value: f16) -> Self { Self::from(value) }
+    fn from_f32(value: f32) -> Self { Self::from(value) }
+    fn from_u32(value: u32) -> Self { Self::from(value) }
+}
+
+
+/// Convert any type into one of the supported sample types.
+/// Should be compiled to a no-op where the file contains the predicted sample type
+pub trait IntoNativeSample: Copy + Default + Sync + 'static {
+
+    /// Convert this sample to an f16, trying to represent the same numerical value.
+    fn to_f16(&self) -> f16;
+
+    /// Convert this sample to an f32, trying to represent the same numerical value.
+    fn to_f32(&self) -> f32;
+
+    /// Convert this sample to an u16, trying to represent the same numerical value.
+    fn to_u32(&self) -> u32;
+}
+
+impl IntoNativeSample for f16 {
+    fn to_f16(&self) -> f16 { f16::from_f16(*self) }
+    fn to_f32(&self) -> f32 { f32::from_f16(*self) }
+    fn to_u32(&self) -> u32 { u32::from_f16(*self) }
+}
+
+impl IntoNativeSample for f32 {
+    fn to_f16(&self) -> f16 { f16::from_f32(*self) }
+    fn to_f32(&self) -> f32 { f32::from_f32(*self) }
+    fn to_u32(&self) -> u32 { u32::from_f32(*self) }
+}
+
+impl IntoNativeSample for u32 {
+    fn to_f16(&self) -> f16 { f16::from_u32(*self) }
+    fn to_f32(&self) -> f32 { f32::from_u32(*self) }
+    fn to_u32(&self) -> u32 { u32::from_u32(*self) }
+}
+
+impl IntoNativeSample for Sample {
+    fn to_f16(&self) -> f16 { Sample::to_f16(*self) }
+    fn to_f32(&self) -> f32 { Sample::to_f32(*self) }
+    fn to_u32(&self) -> u32 { Sample::to_u32(*self) }
+}
+
+
+
